refactor: db into folder for granularity
