+++ src/lib/db/utils.ts
@@ -1,0 +1,8 @@
+import firebase from "../firebase";
+
+export function addNewsletterSubscription(email: string) {
+  return firebase.firestore().collection("subscriptions").doc(email).set({
+    email,
+    createdAt: firebase.firestore.Timestamp.now(),
+  });
+}
